% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSVDdpd.R
\name{cv.alpha}
\alias{cv.alpha}
\title{Calculate optimal robustness parameter}
\usage{
cv.alpha(X, alphas = 10)
}
\arguments{
\item{X}{\code{matrix}, whose singular value decomposition is required}

\item{alphas}{\code{numeric vector}, vector of robustness parameters to try.}
}
\description{
\code{cv.alpha} returns the optimal robustness parameter
}
